// src/lib/sdk/models/DeviceTypeResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeviceTypeResponseDto = {
    /**
     * Device type ID
     */
    id: string;
    /**
     * Device type name
     */
    name: string;
    /**
     * Device type description
     */
    description?: string;
    /**
     * Active status
     */
    isActive: boolean;
    /**
     * Creation timestamp
     */
    createdAt: string;
    /**
     * Last update timestamp
     */
    updatedAt: string;
};